Fix taxonomy filter GUID parsing
